+++ packages/sdk-scripts/jest.config.js
@@ -2,4 +2,5 @@ const { jestConfigFactory } = require("@next-core/jest-config-factory");
 
 module.exports = jestConfigFactory({
   standalone: true,
+  cwd: __dirname,
 });