+++ dll/d3/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.73](https://github.com/easyops-cn/next-core/compare/@next-dll/d3@2.0.72...@next-dll/d3@2.0.73) (2022-11-07)
+
+**Note:** Version bump only for package @next-dll/d3
+
 ## [2.0.72](https://github.com/easyops-cn/next-core/compare/@next-dll/d3@2.0.71...@next-dll/d3@2.0.72) (2022-11-04)
 
 **Note:** Version bump only for package @next-dll/d3