+++ dll/react-dnd/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.471](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.470...@next-dll/react-dnd@0.1.471) (2022-11-11)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.470](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.469...@next-dll/react-dnd@0.1.470) (2022-11-11)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.469](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.468...@next-dll/react-dnd@0.1.469) (2022-11-11)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.468](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.467...@next-dll/react-dnd@0.1.468) (2022-11-10)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.467](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.466...@next-dll/react-dnd@0.1.467) (2022-11-08)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.466](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.465...@next-dll/react-dnd@0.1.466) (2022-11-08)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.465](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.464...@next-dll/react-dnd@0.1.465) (2022-11-08)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.464](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.463...@next-dll/react-dnd@0.1.464) (2022-11-07)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.463](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.462...@next-dll/react-dnd@0.1.463) (2022-11-07)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.462](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.461...@next-dll/react-dnd@0.1.462) (2022-11-07)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.461](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.460...@next-dll/react-dnd@0.1.461) (2022-11-07)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
+## [0.1.460](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.459...@next-dll/react-dnd@0.1.460) (2022-11-04)
+
+**Note:** Version bump only for package @next-dll/react-dnd
+
 ## [0.1.459](https://github.com/easyops-cn/next-core/compare/@next-dll/react-dnd@0.1.458...@next-dll/react-dnd@0.1.459) (2022-11-04)
 
 **Note:** Version bump only for package @next-dll/react-dnd