+++ packages/brick-http/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.6](https://github.com/easyops-cn/next-core/compare/@next-core/brick-http@2.7.5...@next-core/brick-http@2.7.6) (2022-11-07)
+
+**Note:** Version bump only for package @next-core/brick-http
+
 ## [2.7.5](https://github.com/easyops-cn/next-core/compare/@next-core/brick-http@2.7.4...@next-core/brick-http@2.7.5) (2022-11-04)
 
 **Note:** Version bump only for package @next-core/brick-http