+++ packages/brick-scripts/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.5](/compare/@next-core/brick-scripts@3.0.4...@next-core/brick-scripts@3.0.5) (2022-11-07)
+
+**Note:** Version bump only for package @next-core/brick-scripts
+
 ## [3.0.4](/compare/@next-core/brick-scripts@3.0.3...@next-core/brick-scripts@3.0.4) (2022-11-04)
 
 **Note:** Version bump only for package @next-core/brick-scripts