+++ packages/build-config-factory/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.23.56](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.23.55...@next-core/build-config-factory@2.23.56) (2022-11-08)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+## [2.23.55](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.23.54...@next-core/build-config-factory@2.23.55) (2022-11-07)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
+## [2.23.54](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.23.53...@next-core/build-config-factory@2.23.54) (2022-11-04)
+
+**Note:** Version bump only for package @next-core/build-config-factory
+
 ## [2.23.53](https://github.com/easyops-cn/next-core/compare/@next-core/build-config-factory@2.23.52...@next-core/build-config-factory@2.23.53) (2022-11-04)
 
 **Note:** Version bump only for package @next-core/build-config-factory