+++ packages/repo-config/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.4](https://github.com/easyops-cn/next-core/compare/@next-core/repo-config@0.2.3...@next-core/repo-config@0.2.4) (2022-11-07)
+
+**Note:** Version bump only for package @next-core/repo-config
+
 ## [0.2.3](https://github.com/easyops-cn/next-core/compare/@next-core/repo-config@0.2.2...@next-core/repo-config@0.2.3) (2022-11-04)
 
 **Note:** Version bump only for package @next-core/repo-config